FieldMap is now completed
I completed the FieldMap function and transformed your ellipse math in a
function <ellipse overlap> that is called from FieldMap is accurate area
is desired.  But there is problem with this, which is the case when 3 or
more ellipses overlap, which I have no clue how to solve. Need to
discuss this with you.